Introduce separate class for notebook content reading
